requirements.txt
appdirs==1.4.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
black==21.7b0
cffi==1.14.6
cfgv==3.3.0
click==8.0.1
cohesion==1.0.0
distlib==0.3.2
fastapi==0.68.0
filelock==3.0.12
flake8==3.9.2
h11==0.12.0
identify==2.2.13
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
passlib==1.7.4
pathspec==0.9.0
pep517==0.11.0
pip-tools==6.2.0
platformdirs==2.2.0
pre-commit==2.14.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
python-decouple==3.4
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
uvicorn==0.15.0
virtualenv==20.7.2
